fix bfp16 regressions on conv based models

+++ forge/test/models/pytorch/vision/beit/model_utils/utils.py
@@ -2,7 +2,6 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 import requests
-import torch
 from PIL import Image
 from transformers import BeitForImageClassification, BeitImageProcessor
 
@@ -18,4 +17,4 @@ def load_input(variant):
     image = Image.open(requests.get(url, stream=True).raw)
     processor = BeitImageProcessor.from_pretrained(variant)
     inputs = processor(images=image, return_tensors="pt")
-    return [inputs["pixel_values"].to(torch.bfloat16)]
+    return [inputs["pixel_values"]]
