+++ forge/test/models/pytorch/vision/rmbg/model_utils/utils.py
@@ -2,7 +2,6 @@
 
 # SPDX-License-Identifier: Apache-2.0
 import requests
-import torch
 from PIL import Image
 from torchvision import transforms
 from transformers import AutoModelForImageSegmentation
@@ -26,4 +25,4 @@ def load_input():
         ]
     )
     inputs = transform_image(image).unsqueeze(0)
-    return [inputs.to(torch.bfloat16)]
+    return [inputs]
